Updates NY show from Rhonda email 5-10
Wholesale
Rearrange the wholesale section slightly, put catalog and order form at
top of page and form at bottom - Done
I’ll send a new catalog and form - Done
Slider (Main Page)
I’m uploading some new images for the homepage slider section now. If
we could put these before the ones we have now so the new ones will pop
up first. If we need to delete some of the old ones I can tell you
which ones to take down. - added - i didnt take out any old ones, once
we put it live we can see if things are slow (the test site is not
super fast in general)
Shop
When you click on shop, could it open our Etsy shop in another window?
We are on Etsy wholesale now also and I think this is easier than the
Square market place. - yes, take a look and let me know if its doing
what you want
the links to square are now gone...that is good, yes?
I matched up and reordered the links on the "shop" hover to mimic what
is on etsy - that ok
About
delete Paige Rommel. - done
And on the sentence above change dynamic trio to duo.- done
Let me know if it looks ok (spacing and such)
Press
Add a press section. Within that section we’ll have a downloadable
digital press kit - added to the last link (i had to do some spacing
changes, but let me know if it needs changing)
Added the page and some of the links, but let me know what you are
really looking for and I'll make it happen
HiRes
Also need a way for them to access high res images. I was just reading
that on the public section of dropbox you can link to individual files
not a folder … I guess if we list the files on the website then they
could link to the public section of dropbox. Let me know if you can
think of an easier way to do this. - I like this idea - do you have
dropbox or need me to set it up?
After the show I would like to add a stockists section with the stores
that carry our stuff listed. - this one I dont completely follow.  Is
it a new link with the main links or a link in with  "store"?